gave FSISPH its own module
moved FSISPH from sph dir to fsisph dir, parsed out FSISumDensity and interface normal calculation functions
