+++ src/FSISPH/CMakeLists.txt
@@ -1,0 +1,36 @@
+include_directories(.)
+set(FSISPH_inst
+    SolidFSISPHHydroBase
+    FSISPHHydroBase
+    computeSurfaceNormals
+    computeFSISPHSumMassDensity
+    )
+
+
+set(FSISPH_sources )
+
+if(ENABLE_2D)
+  list(APPEND FSISPH_sources
+       FSISPHHydroBaseRZ.cc
+       SolidFSISPHHydroBaseRZ.cc
+      )
+endif()
+
+
+instantiate(FSISPH_inst FSISPH_sources)
+
+set(FSISPH_headers
+    SolidFSISPHHydroBase.hh
+    SolidFSISPHHydroBaseRZ.hh
+    FSISPHHydroBase.hh
+    FSISPHHydroBaseRZ.hh
+    computeSurfaceNormals.hh
+    computeFSISPHSumMassDensity.hh
+    )
+
+spheral_install_python_files(
+  FSISPHHydros.py
+  )
+
+spheral_add_cxx_library(FSISPH)
+
