+++ src/FSISPH/FSISPHHydroBaseRZ.cc
@@ -1,10 +1,17 @@
 //---------------------------------Spheral++----------------------------------//
-// SPHHydroBase -- modified SPHHydro for large density discontinuities
+// SPHHydroBaseRZ -- An SPH/ASPH hydrodynamic package for Spheral++,
+//                   specialized for 2D RZ (cylindrical) geometry.
+//
+// This RZ version is a naive area-weighting implementation, nothing as
+// highfalutin as the Garcia-Senz approach.
+//
+// Note this version is currently abusing our ordinary 2D geometric types,
+// implicitly mapping x->z, y->r.
+//
+// Created by JMO, Fri May  6 16:18:36 PDT 2016
 //----------------------------------------------------------------------------//
 #include "FileIO/FileIO.hh"
-#include "computeSPHSumMassDensity.hh"
-#include "correctSPHSumMassDensity.hh"
-#include "computeSumVoronoiCellMassDensity.hh"
+//#include "computeSumVoronoiCellMassDensity.hh"
 //#include "computeSPHOmegaGradhCorrection.hh"
 #include "NodeList/SmoothingScaleBase.hh"
 #include "Hydro/HydroFieldNames.hh"
@@ -21,7 +28,7 @@
 #include "Hydro/VolumePolicy.hh"
 #include "Hydro/VoronoiMassDensityPolicy.hh"
 #include "Hydro/SumVoronoiMassDensityPolicy.hh"
-#include "Hydro/SpecificThermalEnergyPolicy.hh"
+#include "Hydro/RZNonSymmetricSpecificThermalEnergyPolicy.hh"
 #include "Hydro/SpecificFromTotalThermalEnergyPolicy.hh"
 #include "Hydro/PositionPolicy.hh"
 #include "Hydro/PressurePolicy.hh"
@@ -38,15 +45,10 @@
 #include "Utilities/timingUtilities.hh"
 #include "Utilities/safeInv.hh"
 #include "Utilities/globalBoundingVolumes.hh"
-#include "Utilities/Timer.hh"
 #include "Mesh/Mesh.hh"
 #include "CRKSPH/volumeSpacing.hh"
 
-#include "SPH/FSISPHHydroBase.hh"
-
-#ifdef _OPENMP
-#include "omp.h"
-#endif
+#include "FSISPH/FSISPHHydroBaseRZ.hh"
 
 #include <limits.h>
 #include <float.h>
@@ -55,6 +57,7 @@
 #include <map>
 #include <vector>
 using std::vector;
+using std::map;
 using std::string;
 using std::pair;
 using std::make_pair;
@@ -65,178 +68,161 @@ using std::min;
 using std::max;
 using std::abs;
 
-// Declare timers
-extern Timer TIME_SPH;
-extern Timer TIME_SPHinitializeStartup;
-extern Timer TIME_SPHregister;
-extern Timer TIME_SPHregisterDerivs;
-extern Timer TIME_SPHpreStepInitialize;
-extern Timer TIME_SPHinitialize;
-extern Timer TIME_SPHfinalizeDerivs;
-extern Timer TIME_SPHghostBounds;
-extern Timer TIME_SPHupdateVol;
-extern Timer TIME_SPHenforceBounds;
-extern Timer TIME_SPHevalDerivs;
-extern Timer TIME_SPHevalDerivs_initial;
-extern Timer TIME_SPHevalDerivs_pairs;
-extern Timer TIME_SPHevalDerivs_final;
-
 namespace Spheral {
 
 //------------------------------------------------------------------------------
 // Construct with the given artificial viscosity and kernels.
 //------------------------------------------------------------------------------
-template<typename Dimension>
-FSISPHHydroBase<Dimension>::
-FSISPHHydroBase(const SmoothingScaleBase<Dimension>& smoothingScaleMethod,
-             DataBase<Dimension>& dataBase,
-             ArtificialViscosity<Dimension>& Q,
-             const TableKernel<Dimension>& W,
-             const double filter,
-             const double cfl,
-             const double surfaceForceCoefficient,
-             const double densityStabilizationCoefficient,
-             const double densityDiffusionCoefficient,
-             const double specificThermalEnergyDiffusionCoefficient,
-             const std::vector<int> sumDensityNodeLists,
-             const bool useVelocityMagnitudeForDt,
-             const bool compatibleEnergyEvolution,
-             const bool evolveTotalEnergy,
-             const bool gradhCorrection,
-             const bool XSPH,
-             const bool correctVelocityGradient,
-             const MassDensityType densityUpdate,
-             const HEvolutionType HUpdate,
-             const double epsTensile,
-             const double nTensile,
-             const Vector& xmin,
-             const Vector& xmax):
-  SPHHydroBase<Dimension>(smoothingScaleMethod, 
-                          dataBase,
-                          Q,
-                          W,
-                          W,
-                          filter,
-                          cfl,
-                          useVelocityMagnitudeForDt,
-                          compatibleEnergyEvolution,
-                          evolveTotalEnergy,
-                          gradhCorrection,
-                          XSPH,
-                          correctVelocityGradient,
-                          true,
-                          densityUpdate,
-                          HUpdate,
-                          epsTensile,
-                          nTensile,
-                          xmin,
-                          xmax),
-    mSurfaceForceCoefficient(surfaceForceCoefficient),
-    mDensityStabilizationCoefficient(densityStabilizationCoefficient),
-    mDensityDiffusionCoefficient(densityDiffusionCoefficient),
-    mSpecificThermalEnergyDiffusionCoefficient(specificThermalEnergyDiffusionCoefficient),
-    mSumDensityNodeLists(sumDensityNodeLists){
-  }
+FSISPHHydroBaseRZ::
+FSISPHHydroBaseRZ(const SmoothingScaleBase<Dim<2> >& smoothingScaleMethod,
+               DataBase<Dimension>& dataBase,
+               ArtificialViscosity<Dim<2> >& Q,
+               const TableKernel<Dim<2> >& W,
+               const double filter,
+               const double cfl,
+               const double surfaceForceCoefficient,
+               const double densityStabilizationCoefficient,
+               const double densityDiffusionCoefficient,
+               const double specificThermalEnergyDiffusionCoefficient,
+               const std::vector<int> sumDensityNodeLists,
+               const bool useVelocityMagnitudeForDt,
+               const bool compatibleEnergyEvolution,
+               const bool evolveTotalEnergy,
+               const bool gradhCorrection,
+               const bool XSPH,
+               const bool correctVelocityGradient,
+               const MassDensityType densityUpdate,
+               const HEvolutionType HUpdate,
+               const double epsTensile,
+               const double nTensile,
+               const Vector& xmin,
+               const Vector& xmax):
+  FSISPHHydroBase<Dim<2> >(smoothingScaleMethod,
+                        dataBase,
+                        Q,
+                        W,
+                        filter,
+                        cfl,
+                        surfaceForceCoefficient,
+                        densityStabilizationCoefficient,
+                        densityDiffusionCoefficient,
+                        specificThermalEnergyDiffusionCoefficient,
+                        sumDensityNodeLists,
+                        useVelocityMagnitudeForDt,
+                        compatibleEnergyEvolution,
+                        evolveTotalEnergy,
+                        gradhCorrection,
+                        XSPH,
+                        correctVelocityGradient,
+                        densityUpdate,
+                        HUpdate,
+                        epsTensile,
+                        nTensile,
+                        xmin,
+                        xmax) {
+}
 
 //------------------------------------------------------------------------------
 // Destructor
 //------------------------------------------------------------------------------
-template<typename Dimension>
-FSISPHHydroBase<Dimension>::
-~FSISPHHydroBase() {
+FSISPHHydroBaseRZ::
+~FSISPHHydroBaseRZ() {
 }
 
 //------------------------------------------------------------------------------
-// FSI specialized density summmation
+// On problem start up, we set the RZ flag on the DataBase.
 //------------------------------------------------------------------------------
-template<typename Dimension>
 void
-FSISPHHydroBase<Dimension>::
+FSISPHHydroBaseRZ::
+initializeProblemStartup(DataBase<Dim<2> >& dataBase) {
+  dataBase.isRZ = true;
+  FSISPHHydroBase<Dim<2> >::initializeProblemStartup(dataBase);
+}
+
+//------------------------------------------------------------------------------
+// Register the state we need/are going to evolve.
+//------------------------------------------------------------------------------
+void
+FSISPHHydroBaseRZ::
+registerState(DataBase<Dim<2> >& dataBase,
+              State<Dim<2> >& state) {
+
+  typedef State<Dimension>::PolicyPointer PolicyPointer;
+
+  // The base class does most of it.
+  FSISPHHydroBase<Dim<2> >::registerState(dataBase, state);
+
+  // Are we using the compatible energy evolution scheme?
+  // If so we need to override the ordinary energy registration with a specialized version.
+  if (this->mCompatibleEnergyEvolution) {
+    FieldList<Dimension, Scalar> specificThermalEnergy = dataBase.fluidSpecificThermalEnergy();
+    PolicyPointer thermalEnergyPolicy(new RZNonSymmetricSpecificThermalEnergyPolicy(dataBase));
+    state.enroll(specificThermalEnergy, thermalEnergyPolicy);
+
+    // Get the policy for the position, and add the specific energy as a dependency.
+    PolicyPointer positionPolicy = state.policy(state.buildFieldKey(HydroFieldNames::position, UpdatePolicyBase<Dimension>::wildcard()));
+    positionPolicy->addDependency(HydroFieldNames::specificThermalEnergy);
+  }
+}
+
+//------------------------------------------------------------------------------
+// Finalize the hydro.
+//------------------------------------------------------------------------------
+void
+FSISPHHydroBaseRZ::
 preStepInitialize(const DataBase<Dimension>& dataBase, 
                   State<Dimension>& state,
                   StateDerivatives<Dimension>& derivs) {
 
-  SPHHydroBase<Dimension>::preStepInitialize(dataBase,state,derivs);
-
-  // test to see if any of the nodeLists require their own density sum
-  bool applySelectDensitySum = false;
-  auto numNodeLists = dataBase.numNodeLists();
-  for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-    if (mSumDensityNodeLists[nodeListi]==1){
-      applySelectDensitySum = true;
+  // If we're going to do the SPH summation density, we need to convert the mass
+  // to mass per unit length first.
+  if (densityUpdate() == MassDensityType::RigorousSumDensity or
+      densityUpdate() == MassDensityType::CorrectedSumDensity) {
+    auto       mass = state.fields(HydroFieldNames::mass, 0.0);
+    const auto pos = state.fields(HydroFieldNames::position, Vector::zero);
+    const auto numNodeLists = mass.numFields();
+    for (auto nodeListi = 0u; nodeListi != numNodeLists; ++nodeListi) {
+      const auto n = mass[nodeListi]->numElements();
+      for (auto i = 0u; i < n; ++i) {
+        const auto circi = 2.0*M_PI*abs(pos(nodeListi, i).y());
+        mass(nodeListi, i) /= circi;
+      }
     }
   }
-  //std::cout << applySelectDensitySum;
-  if (applySelectDensitySum){
-      const auto& connectivityMap = dataBase.connectivityMap();
-      const auto& position = state.fields(HydroFieldNames::position, Vector::zero);
-      const auto& mass = state.fields(HydroFieldNames::mass, 0.0);
-      const auto& H = state.fields(HydroFieldNames::H, SymTensor::zero);
-      const auto& W = this->kernel();
-      auto        massDensity = state.fields(HydroFieldNames::massDensity, 0.0);
-      this->computeFSISPHSumMassDensity(connectivityMap, W, position, mass, H, massDensity);
-      for (auto boundaryItr = this->boundaryBegin(); boundaryItr < this->boundaryEnd(); ++boundaryItr) (*boundaryItr)->applyFieldListGhostBoundary(massDensity);
-      for (auto boundaryItr = this->boundaryBegin(); boundaryItr < this->boundaryEnd(); ++boundaryItr) (*boundaryItr)->finalizeGhostBoundary();
+
+  // Base class finalization does most of the work.
+  FSISPHHydroBase<Dimension>::preStepInitialize(dataBase, state, derivs);
+
+  // Now convert back to true masses and mass densities.  We also apply the RZ
+  // correction factor to the mass density.
+  if (densityUpdate() == MassDensityType::RigorousSumDensity or
+      densityUpdate() == MassDensityType::CorrectedSumDensity) {
+    const auto position = state.fields(HydroFieldNames::position, Vector::zero);
+    const auto H = state.fields(HydroFieldNames::H, SymTensor::zero);
+    auto       mass = state.fields(HydroFieldNames::mass, 0.0);
+    auto       massDensity = state.fields(HydroFieldNames::massDensity, 0.0);
+    const auto numNodeLists = massDensity.numFields();
+    for (auto nodeListi = 0u; nodeListi != numNodeLists; ++nodeListi) {
+      const auto n = massDensity[nodeListi]->numElements();
+      for (auto i = 0u; i != n; ++i) {
+        const auto& xi = position(nodeListi, i);
+        const auto  circi = 2.0*M_PI*abs(xi.y());
+        mass(nodeListi, i) *= circi;
+      }
+    }
   }
-
 }
-
-//------------------------------------------------------------------------------
-// Initialize Our AV grad-h moved to eval derivs method
-//------------------------------------------------------------------------------
-// template<typename Dimension>
-// void
-// SPHHydroBase<Dimension>::
-// initialize(const typename Dimension::Scalar time,
-//            const typename Dimension::Scalar dt,
-//            const DataBase<Dimension>& dataBase,
-//            State<Dimension>& state,
-//            StateDerivatives<Dimension>& derivs) {
-//   TIME_SPHinitialize.start();
-
-//   // Initialize the grad h corrrections if needed.
-//   //const TableKernel<Dimension>& W = this->kernel();
-//   const TableKernel<Dimension>& WPi = this->PiKernel();
-
-//   if (mGradhCorrection) {
-//     const ConnectivityMap<Dimension>& connectivityMap = dataBase.connectivityMap();
-//     const FieldList<Dimension, Vector> position = state.fields(HydroFieldNames::position, Vector::zero);
-//     const FieldList<Dimension, SymTensor> H = state.fields(HydroFieldNames::H, SymTensor::zero);
-//     FieldList<Dimension, Scalar> omega = state.fields(HydroFieldNames::omegaGradh, 0.0);
-//     computeSPHOmegaGradhCorrection(connectivityMap, this->kernel(), position, H, omega);
-//     for (ConstBoundaryIterator boundItr = this->boundaryBegin();
-//          boundItr != this->boundaryEnd();
-//          ++boundItr) (*boundItr)->applyFieldListGhostBoundary(omega);
-//   }
-
-//   // Get the artificial viscosity and initialize it.
-//   ArtificialViscosity<Dimension>& Q = this->artificialViscosity();
-//   Q.initialize(dataBase, 
-//                state,
-//                derivs,
-//                this->boundaryBegin(),
-//                this->boundaryEnd(),
-//                time, 
-//                dt,
-//                WPi);
-
-//   // We depend on the caller knowing to finalize the ghost boundaries!
-//   TIME_SPHinitialize.stop();
-// }
 
 //------------------------------------------------------------------------------
 // Determine the principle derivatives.
 //------------------------------------------------------------------------------
-template<typename Dimension>
 void
-FSISPHHydroBase<Dimension>::
-evaluateDerivatives(const typename Dimension::Scalar /*time*/,
-                    const typename Dimension::Scalar /*dt*/,
-                    const DataBase<Dimension>& dataBase,
-                    const State<Dimension>& state,
-                    StateDerivatives<Dimension>& derivatives) const {
-  TIME_SPHevalDerivs.start();
-  TIME_SPHevalDerivs_initial.start();
+FSISPHHydroBaseRZ::
+evaluateDerivatives(const Dim<2>::Scalar /*time*/,
+                    const Dim<2>::Scalar /*dt*/,
+                    const DataBase<Dim<2> >& dataBase,
+                    const State<Dim<2> >& state,
+                    StateDerivatives<Dim<2> >& derivatives) const {
 
   //static double totalLoopTime = 0.0;
   // Get the ArtificialViscosity.
@@ -254,7 +240,7 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
   const auto rhoStabilizeCoeff = this->densityStabilizationCoefficient();
   const auto surfaceForceCoeff = this->surfaceForceCoefficient();
   const auto evolveTotalEnergy = this->evolveTotalEnergy();
-  const auto epsTensile = this->epsilonTensile();
+  //const auto epsTensile = this->epsilonTensile();
   const auto compatibleEnergyEvolution = this->compatibleEnergyEvolution();
   const auto XSPH = this->XSPH();
   const auto& smoothingScaleMethod = this->smoothingScaleMethod();
@@ -279,7 +265,6 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
   CHECK(position.size() == numNodeLists);
   CHECK(velocity.size() == numNodeLists);
   CHECK(massDensity.size() == numNodeLists);
-  CHECK(specificThermalEnergy.size() == numNodeLists);
   CHECK(H.size() == numNodeLists);
   CHECK(pressure.size() == numNodeLists);
   CHECK(soundSpeed.size() == numNodeLists);
@@ -331,16 +316,9 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
   const auto  npairs = pairs.size();
 
   // Size up the pair-wise accelerations before we start.
-  if (compatibleEnergyEvolution) pairAccelerations.resize(npairs);
-
-  // The scale for the tensile correction.
-  const auto& nodeList = mass[0]->nodeList();
-  const auto  nPerh = nodeList.nodesPerSmoothingScale();
-  const auto  WnPerh = W(1.0/nPerh, 1.0);
-  TIME_SPHevalDerivs_initial.stop();
+  if (compatibleEnergyEvolution) pairAccelerations.resize(2*npairs);
 
   // Walk all the interacting pairs.
-  TIME_SPHevalDerivs_pairs.start();
 #pragma omp parallel
   {
     // Thread private scratch variables
@@ -348,12 +326,12 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
     Scalar Wi, gWi, WQi, gWQi, Wj, gWj, WQj, gWQj;
     Tensor QPiij, QPiji;
 
-    typename SpheralThreads<Dimension>::FieldListStack threadStack;
+    typename SpheralThreads<Dim<2>>::FieldListStack threadStack;
     //auto rhoSum_thread = rhoSum.threadCopy(threadStack);
     //auto normalization_thread = normalization.threadCopy(threadStack);
     auto DvDt_thread = DvDt.threadCopy(threadStack);
     auto DepsDt_thread = DepsDt.threadCopy(threadStack);
-    auto DrhoDt_thread = DrhoDt.threadCopy(threadStack);
+    auto DrhoDt_thread = DepsDt.threadCopy(threadStack);
     auto DvDx_thread = DvDx.threadCopy(threadStack);
     //auto localDvDx_thread = localDvDx.threadCopy(threadStack);
     auto M_thread = M.threadCopy(threadStack);
@@ -374,19 +352,22 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       nodeListj = pairs[kk].j_list;
 
       // Get the state for node i.
-      const auto& ri = position(nodeListi, i);
-      const auto& mi = mass(nodeListi, i);
+      const auto& posi = position(nodeListi, i);
+      const auto  ri = abs(posi.y());
+      const auto  circi = 2.0*M_PI*ri;
+      const auto  mi = mass(nodeListi, i);
+      const auto  mRZi = mi/circi;
       const auto& vi = velocity(nodeListi, i);
-      const auto& rhoi = massDensity(nodeListi, i);
-      const auto& epsi = specificThermalEnergy(nodeListi,i);
-      const auto& Pi = pressure(nodeListi, i);
+      const auto  rhoi = massDensity(nodeListi, i);
+      const auto  epsi = specificThermalEnergy(nodeListi, i);
+      const auto  Pi = pressure(nodeListi, i);
       const auto& Hi = H(nodeListi, i);
-      const auto& ci = soundSpeed(nodeListi, i);
+      const auto  ci = soundSpeed(nodeListi, i);
       //const auto& omegai = omega(nodeListi, i);
       const auto  Hdeti = Hi.Determinant();
       //const auto  safeOmegai = safeInv(omegai, tiny);
+      const auto  zetai = abs((Hi*posi).y());
       const auto Ki = max(tiny,rhoi*ci*ci);
-      CHECK(mi > 0.0);
       CHECK(rhoi > 0.0);
       CHECK(Hdeti > 0.0);
 
@@ -408,19 +389,22 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       auto& massSecondMomenti = massSecondMoment_thread(nodeListi, i);
 
       // Get the state for node j
-      const auto& rj = position(nodeListj, j);
-      const auto& mj = mass(nodeListj, j);
+      const auto& posj = position(nodeListj, j);
+      const auto  rj = abs(posj.y());
+      const auto  circj = 2.0*M_PI*rj;
+      const auto  mj = mass(nodeListj, j);
+      const auto  mRZj = mj/circj;
       const auto& vj = velocity(nodeListj, j);
-      const auto& rhoj = massDensity(nodeListj, j);
-      const auto& epsj = specificThermalEnergy(nodeListj,j);
-      const auto& Pj = pressure(nodeListj, j);
+      const auto  rhoj = massDensity(nodeListj, j);
+      const auto  epsj = specificThermalEnergy(nodeListj, j);
+      const auto  Pj = pressure(nodeListj, j);
       const auto& Hj = H(nodeListj, j);
-      const auto& cj = soundSpeed(nodeListj, j);
+      const auto  cj = soundSpeed(nodeListj, j);
       //const auto& omegaj = omega(nodeListj, j);
       const auto  Hdetj = Hj.Determinant();
-      //const auto  safeOmegaj = safeInv(omegaj, tiny); 
+      //const auto  safeOmegaj = safeInv(omegaj, tiny);
+      const auto  zetaj = abs((Hj*posj).y());
       const auto Kj = max(tiny,rhoj*cj*cj);
-      CHECK(mj > 0.0);
       CHECK(rhoj > 0.0);
       CHECK(Hdetj > 0.0);
 
@@ -442,18 +426,17 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       auto& massSecondMomentj = massSecondMoment_thread(nodeListj, j);
 
       // Flag if this is a contiguous material pair or not.
-      const bool sameMatij =  (nodeListi == nodeListj);
+      const bool sameMatij = (nodeListi == nodeListj);
 
       // Node displacement.
-      const auto rij = ri - rj;
+      const auto xij = posi - posj;
       const auto Hij = 0.5*(Hi+Hj);
-      const auto etaij = Hij*rij;
+      const auto etaij = Hij*xij;
       //const auto Hdetij = Hij.Determinant();
       const auto etaMagij = etaij.magnitude();
       CHECK(etaMagij >= 0.0);
-
-      const auto etai = Hi*rij;
-      const auto etaj = Hj*rij;
+      const auto etai = Hi*xij;
+      const auto etaj = Hj*xij;
       const auto etaMagi = etai.magnitude();
       const auto etaMagj = etaj.magnitude();
       CHECK(etaMagi >= 0.0);
@@ -474,20 +457,20 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
 
       // Zero'th and second moment of the node distribution -- used for the
       // ideal H calculation.
-      const auto fweightij = sameMatij ? 1.0 : mj*rhoi/(mi*rhoj);
-      const auto rij2 = rij.magnitude2();
-      const auto thpt = rij.selfdyad()*safeInvVar(rij2*rij2*rij2);
+      const auto fweightij =  1.0;
+      const auto xij2 = xij.magnitude2();
+      const auto thpt = xij.selfdyad()*safeInvVar(xij2*xij2*xij2);
       weightedNeighborSumi +=     fweightij*std::abs(gWi);
       weightedNeighborSumj += 1.0/fweightij*std::abs(gWj);
       massSecondMomenti +=     fweightij*gradWi.magnitude2()*thpt;
       massSecondMomentj += 1.0/fweightij*gradWj.magnitude2()*thpt;
 
       // Contribution to the sum density.
-      //if (sameMatij) {
-      //  rhoSumi += mj*Wi;
-      //  rhoSumj += mi*Wj;
-      //  normi += mi/rhoi*Wi;
-      //  normj += mj/rhoj*Wj;
+      //if (nodeListi == nodeListj) {
+      //  rhoSumi += mRZj*Wi;
+      //  rhoSumj += mRZi*Wj;
+      //  normi += mRZi/rhoi*Wi;
+      //  normj += mRZj/rhoj*Wj;
       //}
 
       // averaged things.
@@ -498,28 +481,29 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       const auto gradWij = 0.5*(gradWi+gradWj);
        
       // volumes
-      const auto voli = mi/rhoi;
-      const auto volj = mj/rhoj;
+      const auto voli = mRZi/rhoi;
+      const auto volj = mRZj/rhoj;
 
       std::tie(QPiij, QPiji) = Q.Piij(nodeListi, i, nodeListj, j,
-                                      ri, etaij, vi, rhoij, cij, Hij,  
-                                      rj, etaij, vj, rhoij, cij, Hij); 
+                                      posi, etaij, vi, rhoij, cij, Hij,  
+                                      posj, etaij, vj, rhoij, cij, Hij); 
 
       //const auto Qacci = 0.5*(QPiij*gradWQi);
       //const auto Qaccj = 0.5*(QPiji*gradWQj);
-      //const auto workQi = 0.5*(QPiij*vij).dot(gradWQi);
-      //const auto workQj = 0.5*(QPiji*vij).dot(gradWQj);
+      // const auto workQi = 0.5*(QPiij*vij).dot(gradWQi);
+      // const auto workQj = 0.5*(QPiji*vij).dot(gradWQj);
       //const auto workQi = vij.dot(Qacci);
       //const auto workQj = vij.dot(Qaccj);
       //const auto Qi = rhoi*rhoi*(QPiij.diagonalElements().maxAbsElement());
       //const auto Qj = rhoj*rhoj*(QPiji.diagonalElements().maxAbsElement());
       //maxViscousPressurei = max(maxViscousPressurei, Qi);
       //maxViscousPressurej = max(maxViscousPressurej, Qj);
-      //effViscousPressurei += mj*Qi*WQi/rhoj;
-      //effViscousPressurej += mi*Qj*WQj/rhoi;
-      //viscousWorki += mj*workQi;
-      //viscousWorkj += mi*workQj;
+      //effViscousPressurei += mRZj*Qi*WQi/rhoj;
+      //effViscousPressurej += mRZi*Qj*WQj/rhoi;
+      //viscousWorki += mRZj*workQi;
+      //viscousWorkj += mRZi*workQj;
 
+    
       // Nominal Velocity Gradient
       const auto vij = vi-vj; 
       auto deltaDvDxi = vij.dyad(gradWi);
@@ -527,13 +511,11 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
 
       // Determine an effective pressure including a term to fight the tensile instability.
       // const auto fij = epsTensile*pow(Wi/(Hdeti*WnPerh), nTensile);
-      const auto fij = epsTensile*FastMath::pow4(Wi/(Hdeti*WnPerh));
-      const auto Ri = fij*(Pi < 0.0 ? -Pi : 0.0);
-      const auto Rj = fij*(Pj < 0.0 ? -Pj : 0.0);
+      //const auto fij = epsTensile*FastMath::pow4(Wi/(Hdeti*WnPerh));
+      //const auto Ri = fij*(Pi < 0.0 ? -Pi : 0.0);
+      //const auto Rj = fij*(Pj < 0.0 ? -Pj : 0.0);
 
       // Pressure
-      const auto Peffi = Pi + Ri;
-      const auto Peffj = Pj + Rj;
       const auto oneOverRhoiRhoj = 1.0/(rhoi*rhoj);
       auto Prhoi = QPiij.Trace()/6.0;
       auto Prhoj = QPiji.Trace()/6.0;
@@ -542,8 +524,8 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       if (sameMatij){
 
         // effective pressure
-        Prhoi += Peffi*oneOverRhoiRhoj;
-        Prhoj += Peffj*oneOverRhoiRhoj;
+        Prhoi += Pi*oneOverRhoiRhoj;
+        Prhoj += Pj*oneOverRhoiRhoj;
 
         // optional density diffusion
         if (rhoDiffusionCoeff>tiny){
@@ -555,24 +537,24 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
         // optional specific thermal energy diffusion
         if (epsDiffusionCoeff>tiny){
           const auto diffusion = epsDiffusionCoeff*(epsi-epsj)*cij/rhoij*etaij.dot(gradWij)/(etaMagij*etaMagij+tiny);
-          DepsDti += mj*diffusion;
-          DepsDtj -= mi*diffusion;
+          DepsDti += mRZj*diffusion;
+          DepsDtj -= mRZi*diffusion;
         }
 
       // different material
       }else{
-
-        const auto isExpanding = vij.dot(rij) > 0; 
+        
+        const auto isExpanding = vij.dot(xij) > 0; 
 
         // interface pressure
-        const auto Pstar = (Peffi*rhoj+Peffj*rhoi)/(rhoi+rhoj);
+        const auto Pstar = (Pi*rhoj+Pj*rhoi)/(rhoi+rhoj);
         const auto sf =  1.0 + surfaceForceCoeff*abs((rhoi-rhoj)/(rhoi+rhoj));
         Prhoi += sf*Pstar*oneOverRhoiRhoj;
         Prhoj += sf*Pstar*oneOverRhoiRhoj;
 
         // Velocity gradient.
-        auto yi =  1.0 + (Pstar-Peffi)/(Ki);
-        auto yj =  1.0 + (Pstar-Peffj)/(Kj);
+        auto yi =  1.0 + (Pstar-Pi)/(Ki);
+        auto yj =  1.0 + (Pstar-Pj)/(Kj);
         if (isExpanding){
           const auto tempVar = yi;
           yi=yj;
@@ -590,11 +572,14 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       
       // Eqn of Motion
       const auto deltaDvDt = Prhoi*gradWi + Prhoj*gradWj;
-      DvDti -= mj*deltaDvDt;
-      DvDtj += mi*deltaDvDt;
+      DvDti -= mRZj*deltaDvDt;
+      DvDtj += mRZi*deltaDvDt;
 
-      if (compatibleEnergyEvolution) pairAccelerations[kk] = -mj*deltaDvDt;  // Acceleration for i (j anti-symmetric)
-      
+      if (compatibleEnergyEvolution) {
+        pairAccelerations[2*kk]   = -mRZj*deltaDvDt;
+        pairAccelerations[2*kk+1] =  mRZi*deltaDvDt;
+      }
+
       // Finish up Velocity Gradient
       if (rhoStabilizeCoeff>tiny){
         const auto diffusion = rhoStabilizeCoeff*cij*etaij.dot(gradWij)/(etaMagij*etaMagij+tiny);
@@ -608,17 +593,17 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       DvDxj -= voli*deltaDvDxj;
 
     // Energy Conservation
-      DepsDti += mj*(Prhoi*deltaDvDxi.Trace());
-      DepsDtj += mi*(Prhoj*deltaDvDxj.Trace());
+      DepsDti += mRZj*(Prhoi*deltaDvDxi.Trace());
+      DepsDtj += mRZi*(Prhoj*deltaDvDxj.Trace());
 
       // Linear gradient correction term.
       if(this->mCorrectVelocityGradient){
-        Mi -= volj*rij.dyad(gradWi);
-        Mj -= voli*rij.dyad(gradWj);
+        Mi -= volj*xij.dyad(gradWi);
+        Mj -= voli*xij.dyad(gradWj);
       }
 
       // Estimate of delta v (for XSPH).
-      if (XSPH and (sameMatij)) {
+      if (XSPH and (sameMatij or min(zetai, zetaj) < 1.0)) {
         const auto wXSPHij = 0.5*(voli*Wi + volj*Wj);
         XSPHWeightSumi += wXSPHij;
         XSPHWeightSumj += wXSPHij;
@@ -626,16 +611,16 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
         XSPHDeltaVj += wXSPHij*vij;
       }
 
+
     } // loop over pairs
 
     // Reduce the thread values to the master.
     threadReduceFieldLists<Dimension>(threadStack);
 
   }   // OpenMP parallel region
-  TIME_SPHevalDerivs_pairs.stop();
+
 
   // Finish up the derivatives for each point.
-  TIME_SPHevalDerivs_final.start();
   for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
     const auto& nodeList = mass[nodeListi]->nodeList();
     const auto  hmin = nodeList.hmin();
@@ -648,14 +633,20 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
     for (auto i = 0u; i < ni; ++i) {
 
       // Get the state for node i.
-      const auto& ri = position(nodeListi, i);
-      const auto& mi = mass(nodeListi, i);
+      const auto& posi = position(nodeListi, i);
+      const auto  ri = abs(posi.y());
+      const auto  circi = 2.0*M_PI*ri;
+      const auto  mi = mass(nodeListi, i);
+      const auto  mRZi = mi/circi;
       const auto& vi = velocity(nodeListi, i);
-      const auto& rhoi = massDensity(nodeListi, i);
+      const auto  rhoi = massDensity(nodeListi, i);
+      const auto  Pi = pressure(nodeListi, i);
       const auto& Hi = H(nodeListi, i);
       const auto  Hdeti = Hi.Determinant();
+      const auto  zetai = abs((Hi*posi).y());
+      const auto  hri = ri*safeInv(zetai);
+      const auto  riInv = safeInv(ri, 0.25*hri);
       const auto  numNeighborsi = connectivityMap.numNeighborsForNode(nodeListi, i);
-      CHECK(mi > 0.0);
       CHECK(rhoi > 0.0);
       CHECK(Hdeti > 0.0);
 
@@ -668,6 +659,7 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       auto& DvDxi = DvDx(nodeListi, i);
       //auto& localDvDxi = localDvDx(nodeListi, i);
       auto& Mi = M(nodeListi, i);
+      //auto& localMi = localM(nodeListi, i);
       auto& DHDti = DHDt(nodeListi, i);
       auto& Hideali = Hideal(nodeListi, i);
       auto& XSPHWeightSumi = XSPHWeightSum(nodeListi, i);
@@ -676,19 +668,28 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
       auto& massSecondMomenti = massSecondMoment(nodeListi, i);
 
       // Add the self-contribution to density sum.
-      //rhoSumi += mi*W0*Hdeti;
-      //normi += mi/rhoi*W0*Hdeti;
+      //rhoSumi += mRZi*W0*Hdeti;
+      //rhoSumi /= circi;
+      //normi += mRZi/rhoi*W0*Hdeti;
 
       // Finish the gradient of the velocity.
+      CHECK(rhoi > 0.0);
       if (this->mCorrectVelocityGradient and
           std::abs(Mi.Determinant()) > 1.0e-10 and
           numNeighborsi > Dimension::pownu(2)) {
         Mi = Mi.Inverse();
         DvDxi = DvDxi*Mi;
-      } 
+      }
 
-      // Evaluate the continuity equation.
-      DrhoDti -=  rhoi*DvDxi.Trace();
+      // Finish the continuity equation.
+      XSPHWeightSumi += Hdeti*mRZi/rhoi*W0;
+      CHECK2(XSPHWeightSumi != 0.0, i << " " << XSPHWeightSumi);
+      XSPHDeltaVi /= XSPHWeightSumi;
+      const auto vri = vi.y(); // + XSPHDeltaVi.y();
+      DrhoDti -= rhoi*(DvDxi.Trace() + vri*riInv);
+
+      // Finish the specific thermal energy evolution.
+      DepsDti -= Pi/rhoi*vri*riInv;
 
       // If needed finish the total energy derivative.
       if (evolveTotalEnergy) DepsDti = mi*(vi.dot(DvDti) + DepsDti);
@@ -699,23 +700,21 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
 
       // Determine the position evolution, based on whether we're doing XSPH or not.
       if (XSPH) {
-        XSPHWeightSumi += Hdeti*mi/rhoi*W0;
-        CHECK2(XSPHWeightSumi != 0.0, i << " " << XSPHWeightSumi);
-        DxDti = vi + XSPHDeltaVi/max(tiny, XSPHWeightSumi);
+        DxDti = vi + XSPHDeltaVi;
       } else {
         DxDti = vi;
       }
 
       // The H tensor evolution.
       DHDti = smoothingScaleMethod.smoothingScaleDerivative(Hi,
-                                                             ri,
+                                                             posi,
                                                              DvDxi,
                                                              hmin,
                                                              hmax,
                                                              hminratio,
                                                              nPerh);
       Hideali = smoothingScaleMethod.newSmoothingScale(Hi,
-                                                        ri,
+                                                        posi,
                                                         weightedNeighborSumi,
                                                         massSecondMomenti,
                                                         W,
@@ -728,107 +727,81 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
                                                         i);
     }
   }
-  TIME_SPHevalDerivs_final.stop();
-  TIME_SPHevalDerivs.stop();
 }
 
-
-
-//===================================================================================
-// Allows density sum to only be applied to some nodeLists
-//===================================================================================
-template<typename Dimension>
+//------------------------------------------------------------------------------
+// Apply the ghost boundary conditions for hydro state fields.
+//------------------------------------------------------------------------------
 void
-FSISPHHydroBase<Dimension>::
-computeFSISPHSumMassDensity(const ConnectivityMap<Dimension>& connectivityMap,
-                            const TableKernel<Dimension>& W,
-                            const FieldList<Dimension, typename Dimension::Vector>& position,
-                            const FieldList<Dimension, typename Dimension::Scalar>& mass,
-                            const FieldList<Dimension, typename Dimension::SymTensor>& H,
-                            FieldList<Dimension, typename Dimension::Scalar>& massDensity) {
+FSISPHHydroBaseRZ::
+applyGhostBoundaries(State<Dim<2> >& state,
+                     StateDerivatives<Dim<2> >& derivs) {
 
-  // Pre-conditions.
-  const auto numNodeLists = massDensity.size();
-  REQUIRE(position.size() == numNodeLists);
-  REQUIRE(mass.size() == numNodeLists);
-  REQUIRE(H.size() == numNodeLists);
-
-  // Some useful variables.
-  const auto W0 = W.kernelValue(0.0, 1.0);
-
-  //typedef typename Dimension::Scalar Scalar;
-  //FieldList<Dimension, Scalar> storeDensity(FieldStorageType::CopyFields);
-  //for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-  //  storeDensity.appendNewField("storeDensity", massDensity[nodeListi]->nodeList(), 0.0);
-  //}
-
-    // First the self contribution.
-  for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-    const auto n = massDensity[nodeListi]->numInternalElements();
-    if (mSumDensityNodeLists[nodeListi]==1){
-#pragma omp parallel for
-      for (auto i = 0u; i < n; ++i) {
-        const auto  mi = mass(nodeListi, i);
-        const auto& Hi = H(nodeListi, i);
-        const auto  Hdeti = Hi.Determinant();
-        //const auto rhoi = storeDensity(nodeListi, i);
-        massDensity(nodeListi,i) = mi*Hdeti*W0;
-      }
-    }
-  }
-  // The set of interacting node pairs.
-  const auto& pairs = connectivityMap.nodePairList();
-  const auto  npairs = pairs.size();
-
-  // Now the pair contributions.
-#pragma omp parallel
-  {
-    int i, j, nodeListi, nodeListj;
-    //auto storeMassDensity_thread = storeDensity.threadCopy();
-    auto massDensity_thread = massDensity.threadCopy();
-
-#pragma omp for
-    for (auto k = 0u; k < npairs; ++k) {
-      i = pairs[k].i_node;
-      j = pairs[k].j_node;
-      nodeListi = pairs[k].i_list;
-      nodeListj = pairs[k].j_list;
-
-      if(mSumDensityNodeLists[nodeListi]==1 || mSumDensityNodeLists[nodeListj]==1){
-        // State for node i
-        const auto& ri = position(nodeListi, i);
-        const auto  mi = mass(nodeListi, i);
-        const auto& Hi = H(nodeListi, i);
-        const auto  Hdeti = Hi.Determinant();
-      
-        // State for node j
-        const auto& rj = position(nodeListj, j);
-        const auto  mj = mass(nodeListj, j);
-        const auto& Hj = H(nodeListj, j);
-        const auto  Hdetj = Hj.Determinant();
-      
-        // Kernel weighting and gradient.
-        const auto rij = ri - rj;
-        const auto etai = (Hi*rij).magnitude();
-        const auto etaj = (Hj*rij).magnitude();
-        const auto Wi = W.kernelValue(etai, Hdeti);
-        const auto Wj = W.kernelValue(etaj, Hdetj);
-
-        massDensity_thread(nodeListi, i) += (mSumDensityNodeLists[nodeListi]==1 ? 
-                                            (nodeListi == nodeListj ? mj : mi)*Wi : 
-                                             0.0);
-        massDensity_thread(nodeListj, j) += (mSumDensityNodeLists[nodeListj]==1 ? 
-                                            (nodeListi == nodeListj ? mi : mj)*Wj : 
-                                             0.0);
-      }
-    }
-
-#pragma omp critical
-    {
-      massDensity_thread.threadReduce();
+  // Convert the mass to mass/length before BCs are applied.
+  FieldList<Dimension, Scalar> mass = state.fields(HydroFieldNames::mass, 0.0);
+  const FieldList<Dimension, Vector> pos = state.fields(HydroFieldNames::position, Vector::zero);
+  const unsigned numNodeLists = mass.numFields();
+  for (unsigned nodeListi = 0; nodeListi != numNodeLists; ++nodeListi) {
+    const unsigned n = mass[nodeListi]->numElements();
+    for (unsigned i = 0; i != n; ++i) {
+      const Scalar circi = 2.0*M_PI*abs(pos(nodeListi, i).y());
+      CHECK(circi > 0.0);
+      mass(nodeListi, i) /= circi;
     }
   }
 
+  // Apply ordinary SPH BCs.
+  FSISPHHydroBase<Dim<2> >::applyGhostBoundaries(state, derivs);
+  for (ConstBoundaryIterator boundItr = this->boundaryBegin();
+       boundItr != this->boundaryEnd();
+       ++boundItr) (*boundItr)->finalizeGhostBoundary();
+
+  // Scale back to mass.
+  for (unsigned nodeListi = 0; nodeListi != numNodeLists; ++nodeListi) {
+    const unsigned n = mass[nodeListi]->numElements();
+    for (unsigned i = 0; i != n; ++i) {
+      const Scalar circi = 2.0*M_PI*abs(pos(nodeListi, i).y());
+      CHECK(circi > 0.0);
+      mass(nodeListi, i) *= circi;
+    }
+  }
+}
+
+//------------------------------------------------------------------------------
+// Enforce the boundary conditions for hydro state fields.
+//------------------------------------------------------------------------------
+void
+FSISPHHydroBaseRZ::
+enforceBoundaries(State<Dim<2> >& state,
+                  StateDerivatives<Dim<2> >& derivs) {
+
+  // Convert the mass to mass/length before BCs are applied.
+  FieldList<Dimension, Scalar> mass = state.fields(HydroFieldNames::mass, 0.0);
+  FieldList<Dimension, Vector> pos = state.fields(HydroFieldNames::position, Vector::zero);
+  const unsigned numNodeLists = mass.numFields();
+  for (unsigned nodeListi = 0; nodeListi != numNodeLists; ++nodeListi) {
+    const unsigned n = mass[nodeListi]->numElements();
+    for (unsigned i = 0; i != n; ++i) {
+      const Scalar circi = 2.0*M_PI*abs(pos(nodeListi, i).y());
+      CHECK(circi > 0.0);
+      mass(nodeListi, i) /= circi;
+    }
+  }
+
+  // Apply ordinary SPH BCs.
+  FSISPHHydroBase<Dim<2> >::enforceBoundaries(state, derivs);
+
+  // Scale back to mass.
+  // We also ensure no point approaches the z-axis too closely.
+  FieldList<Dimension, SymTensor> H = state.fields(HydroFieldNames::H, SymTensor::zero);
+  for (unsigned nodeListi = 0; nodeListi != numNodeLists; ++nodeListi) {
+    const unsigned n = mass[nodeListi]->numElements();
+    for (unsigned i = 0; i != n; ++i) {
+      Vector& posi = pos(nodeListi, i);
+      const Scalar circi = 2.0*M_PI*abs(posi.y());
+      mass(nodeListi, i) *= circi;
+    }
+  }
 }
 
 }
