+++ src/FSISPH/SolidFSISPHHydroBase.cc
@@ -38,7 +38,9 @@
 #include "SolidMaterial/SolidEquationOfState.hh"
 #include "Utilities/Timer.hh"
 
-#include "SolidFSISPHHydroBase.hh"
+#include "FSISPH/SolidFSISPHHydroBase.hh"
+#include "FSISPH/computeSurfaceNormals.hh"
+#include "FSISPH/computeFSISPHSumMassDensity.hh"
 
 #include <limits.h>
 #include <float.h>
@@ -58,6 +60,8 @@ using std::min;
 using std::max;
 using std::abs;
 
+extern Timer TIME_SolidFSISPHpreStepInitialize;
+extern Timer TIME_SolidFSISPHinitialize;
 extern Timer TIME_SolidFSISPHregisterDerivs;
 extern Timer TIME_SolidFSISPHregisterState;
 
@@ -160,9 +164,19 @@ SolidFSISPHHydroBase(const SmoothingScaleBase<Dimension>& smoothingScaleMethod,
   mDensityStabilizationCoefficient(densityStabilizationCoefficient),
   mDensityDiffusionCoefficient(densityDiffusionCoefficient),
   mSpecificThermalEnergyDiffusionCoefficient(specificThermalEnergyDiffusionCoefficient),
+  mApplySelectDensitySum(false),
   mSumDensityNodeLists(sumDensityNodeLists),
   mSurfaceNormals(FieldStorageType::CopyFields){
      mSurfaceNormals = dataBase.newSolidFieldList(Vector::zero,  "FSISurfaceNormals");
+
+    // see if we're summing density for any nodelist
+    auto numNodeLists = dataBase.numNodeLists();
+    for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
+      if (sumDensityNodeLists[nodeListi]==1){
+        mApplySelectDensitySum = true;
+      } 
+    }
+    
   }
 
 //------------------------------------------------------------------------------
@@ -203,8 +217,8 @@ registerState(DataBase<Dimension>& dataBase,
 template<typename Dimension>
 void
 SolidFSISPHHydroBase<Dimension>::
-registerDerivatives(DataBase<Dimension>& dataBase,
-                    StateDerivatives<Dimension>& derivs) {
+registerDerivatives(DataBase<Dimension>&  /*dataBase*/,
+                    StateDerivatives<Dimension>& /*derivs*/) {
   TIME_SolidFSISPHregisterDerivs.start();
 
   // Call the ancestor method.
@@ -222,19 +236,8 @@ SolidFSISPHHydroBase<Dimension>::
 preStepInitialize(const DataBase<Dimension>& dataBase, 
                   State<Dimension>& state,
                   StateDerivatives<Dimension>& derivs) {
-
-  SolidSPHHydroBase<Dimension>::preStepInitialize(dataBase,state,derivs);
-
-  // test to see if any of the nodeLists require their own density sum
-  bool applySelectDensitySum = false;
-  auto numNodeLists = dataBase.numNodeLists();
-  for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-    if (mSumDensityNodeLists[nodeListi]==1){
-      applySelectDensitySum = true;
-    }
-  }
-  //std::cout << applySelectDensitySum;
-  if (applySelectDensitySum){
+  TIME_SolidFSISPHpreStepInitialize.start();
+  if (mApplySelectDensitySum){
       const auto& connectivityMap = dataBase.connectivityMap();
       const auto& position = state.fields(HydroFieldNames::position, Vector::zero);
       const auto& mass = state.fields(HydroFieldNames::mass, 0.0);
@@ -243,11 +246,11 @@ preStepInitialize(const DataBase<Dimension>& dataBase,
       const auto& H = state.fields(HydroFieldNames::H, SymTensor::zero);
       const auto& W = this->kernel();
       auto        massDensity = state.fields(HydroFieldNames::massDensity, 0.0);
-      this->computeFSISPHSumMassDensity(connectivityMap, W, position, mass, H, massDensity);
+      computeFSISPHSumMassDensity(connectivityMap, W, mSumDensityNodeLists, position, mass, H, massDensity);
       for (auto boundaryItr = this->boundaryBegin(); boundaryItr < this->boundaryEnd(); ++boundaryItr) (*boundaryItr)->applyFieldListGhostBoundary(massDensity);
       for (auto boundaryItr = this->boundaryBegin(); boundaryItr < this->boundaryEnd(); ++boundaryItr) (*boundaryItr)->finalizeGhostBoundary();
   }
-
+  TIME_SolidFSISPHpreStepInitialize.stop();
 }
 
 
@@ -263,7 +266,7 @@ initialize(const typename Dimension::Scalar time,
            const DataBase<Dimension>& dataBase,
            State<Dimension>& state,
            StateDerivatives<Dimension>& derivs) {
-  //TIME_SPHinitialize.start();
+  TIME_SolidFSISPHinitialize.start();
 
   const TableKernel<Dimension>& W = this->kernel();
 
@@ -277,33 +280,35 @@ initialize(const typename Dimension::Scalar time,
                dt,
                W);
 
-  const auto& connectivityMap = dataBase.connectivityMap();
-  const auto& position = state.fields(HydroFieldNames::position, Vector::zero);
-  const auto& mass = state.fields(HydroFieldNames::mass, 0.0);
-  const auto& massDensity = state.fields(HydroFieldNames::massDensity, 0.0);
-  const auto& H = state.fields(HydroFieldNames::H, SymTensor::zero);
-        auto& normals = this->mSurfaceNormals;
-  normals.Zero();
-  this->computeSurfaceNormals(connectivityMap,
-                              W,
-                              position,
-                              mass,
-                              massDensity,
-                              H,
-                              normals);
+  // put some trigger for this in Slip interface bc?
+  if (true){  
+    const auto& connectivityMap = dataBase.connectivityMap();
+    const auto& position = state.fields(HydroFieldNames::position, Vector::zero);
+    const auto& mass = state.fields(HydroFieldNames::mass, 0.0);
+    const auto& massDensity = state.fields(HydroFieldNames::massDensity, 0.0);
+    const auto& H = state.fields(HydroFieldNames::H, SymTensor::zero);
+          auto& normals = this->mSurfaceNormals;
+    normals.Zero();
+    computeSurfaceNormals(connectivityMap,
+                          W,
+                          position,
+                          mass,
+                          massDensity,
+                          H,
+                          normals);
 
 
-  for (ConstBoundaryIterator boundaryItr = this->boundaryBegin();
-       boundaryItr != this->boundaryEnd();
-       ++boundaryItr) {
-      (*boundaryItr)->applyFieldListGhostBoundary(normals);
-  }
-  for (ConstBoundaryIterator boundaryItr = this->boundaryBegin(); 
+    for (ConstBoundaryIterator boundaryItr = this->boundaryBegin();
          boundaryItr != this->boundaryEnd();
-         ++boundaryItr) (*boundaryItr)->finalizeGhostBoundary();
-
+         ++boundaryItr) {
+        (*boundaryItr)->applyFieldListGhostBoundary(normals);
+    }
+    for (ConstBoundaryIterator boundaryItr = this->boundaryBegin(); 
+           boundaryItr != this->boundaryEnd();
+           ++boundaryItr) (*boundaryItr)->finalizeGhostBoundary();
+  }
   // We depend on the caller knowing to finalize the ghost boundaries!
-  //TIME_SPHinitialize.stop();
+  TIME_SolidFSISPHinitialize.stop();
 }
 
 
@@ -934,15 +939,12 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
        DepsDtj -= mi*diffusion;
       }
 
-      //if (sameMatij){
-      const auto diffusion = (Si-Sj)*cij*etaij.dot(gradWij)/(etaMagij*etaMagij+tiny);
-      DSDti += volj*diffusion;
-      DSDtj -= voli*diffusion;
-      //}else{
-       //const auto diffusion =  cij*etaij.dot(gradWij)/(etaMagij*etaMagij+tiny);
-       //DSDti += volj*Si*diffusion;
-       //DSDtj += voli*Sj*diffusion; 
-      //}
+      // rigorous enforcement of single-valued stress-state at interface
+      if (!sameMatij){
+        const auto diffusion = (Si-Sj)*cij*etaij.dot(gradWij)/(etaMagij*etaMagij+tiny);
+        DSDti += volj*diffusion;
+        DSDtj -= voli*diffusion;
+      }
 
       // XSPH
       //-----------------------------------------------------------
@@ -1073,194 +1075,6 @@ evaluateDerivatives(const typename Dimension::Scalar /*time*/,
     }
   }
 }
-
-
-//===================================================================================
-// Allows density sum to only be applied to some nodeLists
-//===================================================================================
-template<typename Dimension>
-void
-SolidFSISPHHydroBase<Dimension>::
-computeFSISPHSumMassDensity(const ConnectivityMap<Dimension>& connectivityMap,
-                            const TableKernel<Dimension>& W,
-                            const FieldList<Dimension, typename Dimension::Vector>& position,
-                            const FieldList<Dimension, typename Dimension::Scalar>& mass,
-                            const FieldList<Dimension, typename Dimension::SymTensor>& H,
-                            FieldList<Dimension, typename Dimension::Scalar>& massDensity) {
-
-  // Pre-conditions.
-  const auto numNodeLists = massDensity.size();
-  REQUIRE(position.size() == numNodeLists);
-  REQUIRE(mass.size() == numNodeLists);
-  REQUIRE(H.size() == numNodeLists);
-
-  // Some useful variables.
-  const auto W0 = W.kernelValue(0.0, 1.0);
-
-  typedef typename Dimension::Scalar Scalar;
-  FieldList<Dimension, Scalar> storeDensity(FieldStorageType::CopyFields);
-  for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-    storeDensity.appendNewField("storeDensity", massDensity[nodeListi]->nodeList(), 0.0);
-  }
-
-  // The set of interacting node pairs.
-  const auto& pairs = connectivityMap.nodePairList();
-  const auto  npairs = pairs.size();
-
-  // Now the pair contributions.
-#pragma omp parallel
-  {
-    int i, j, nodeListi, nodeListj;
-    auto storeMassDensity_thread = storeDensity.threadCopy();
-
-#pragma omp for
-    for (auto k = 0u; k < npairs; ++k) {
-      i = pairs[k].i_node;
-      j = pairs[k].j_node;
-      nodeListi = pairs[k].i_list;
-      nodeListj = pairs[k].j_list;
-
-      if(mSumDensityNodeLists[nodeListi]==1 || mSumDensityNodeLists[nodeListj]==1){
-        // State for node i
-        const auto& ri = position(nodeListi, i);
-        const auto  mi = mass(nodeListi, i);
-        const auto& Hi = H(nodeListi, i);
-        const auto  Hdeti = Hi.Determinant();
-      
-        // State for node j
-        const auto& rj = position(nodeListj, j);
-        const auto  mj = mass(nodeListj, j);
-        const auto& Hj = H(nodeListj, j);
-        const auto  Hdetj = Hj.Determinant();
-      
-        // Kernel weighting and gradient.
-        const auto rij = ri - rj;
-        const auto etai = (Hi*rij).magnitude();
-        const auto etaj = (Hj*rij).magnitude();
-        const auto Wi = W.kernelValue(etai, Hdeti);
-        const auto Wj = W.kernelValue(etaj, Hdetj);
-
-        storeMassDensity_thread(nodeListi, i) += (mSumDensityNodeLists[nodeListi]==1 ? 
-                                            (nodeListi == nodeListj ? mj : mi)*Wi : 
-                                             0.0);
-        storeMassDensity_thread(nodeListj, j) += (mSumDensityNodeLists[nodeListj]==1 ? 
-                                            (nodeListi == nodeListj ? mi : mj)*Wj : 
-                                             0.0);
-      }
-    }
-
-#pragma omp critical
-    {
-      storeMassDensity_thread.threadReduce();
-    }
-  }
-
-  for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-    const auto n = massDensity[nodeListi]->numInternalElements();
-    if (mSumDensityNodeLists[nodeListi]==1){
-#pragma omp parallel for
-      for (auto i = 0u; i < n; ++i) {
-        const auto  mi = mass(nodeListi, i);
-        const auto& Hi = H(nodeListi, i);
-        const auto  Hdeti = Hi.Determinant();
-        massDensity(nodeListi,i) = storeDensity(nodeListi, i) + mi*Hdeti*W0;
-      }
-    }
-  }
-
-}
-
-
-
-//==========================================================================
-// Unit normals for interfaces
-//==========================================================================
-template<typename Dimension>
-void
-SolidFSISPHHydroBase<Dimension>::
-computeSurfaceNormals(const ConnectivityMap<Dimension>& connectivityMap,
-                            const TableKernel<Dimension>& W,
-                            const FieldList<Dimension, typename Dimension::Vector>& position,
-                            const FieldList<Dimension, typename Dimension::Scalar>& mass,
-                            const FieldList<Dimension, typename Dimension::Scalar>& massDensity,
-                            const FieldList<Dimension, typename Dimension::SymTensor>& H,
-                            FieldList<Dimension, typename Dimension::Vector>& interfaceNormals) {
-
-  // Pre-conditions.
-  const auto numNodeLists = massDensity.size();
-  REQUIRE(position.size() == numNodeLists);
-  REQUIRE(mass.size() == numNodeLists);
-  REQUIRE(H.size() == numNodeLists);
-
-  // The set of interacting node pairs.
-  const auto& pairs = connectivityMap.nodePairList();
-  const auto  npairs = pairs.size();
-
-  // Now the pair contributions.
-#pragma omp parallel
-  {
-    int i, j, nodeListi, nodeListj;
-    auto interfaceNormals_thread = interfaceNormals.threadCopy();
-
-#pragma omp for
-    for (auto k = 0u; k < npairs; ++k) {
-      i = pairs[k].i_node;
-      j = pairs[k].j_node;
-      nodeListi = pairs[k].i_list;
-      nodeListj = pairs[k].j_list;
-
-      if(nodeListi!=nodeListj){
-        // State for node i
-        const auto& ri = position(nodeListi, i);
-        const auto  mi = mass(nodeListi, i);
-        const auto  rhoi = massDensity(nodeListi, i);
-        const auto& Hi = H(nodeListi, i);
-        const auto  Hdeti = Hi.Determinant();
-      
-        // State for node j
-        const auto& rj = position(nodeListj, j);
-        const auto  mj = mass(nodeListj, j);
-        const auto  rhoj = massDensity(nodeListj, j);
-        const auto& Hj = H(nodeListj, j);
-        const auto  Hdetj = Hj.Determinant();
-      
-        // Kernel weighting and gradient.
-        const auto rij = ri - rj;
-        const auto etai = Hi*rij;
-        const auto etaj = Hj*rij;
-        const auto etaMagi = etai.magnitude();
-        const auto etaMagj = etaj.magnitude();
-        const auto Hetai = Hi*etai.unitVector();
-        const auto Hetaj = Hj*etaj.unitVector();
-
-        const auto gWi = W.gradValue(etaMagi, Hdeti);
-        auto gradWi = gWi*Hetai;
-
-        const auto gWj = W.gradValue(etaMagj, Hdetj);
-        auto gradWj = gWj*Hetaj;
-
-        interfaceNormals_thread(nodeListi, i) += mj/rhoj  * gradWi;
-        interfaceNormals_thread(nodeListj, j) -= mi/rhoi  * gradWj;
-      }
-    }
-
-#pragma omp critical
-    {
-      interfaceNormals_thread.threadReduce();
-    }
-  }
-
-  for (auto nodeListi = 0u; nodeListi < numNodeLists; ++nodeListi) {
-     const auto n = interfaceNormals[nodeListi]->numInternalElements();
- #pragma omp parallel for
-     for (auto i = 0u; i < n; ++i) {
-       interfaceNormals(nodeListi,i) /= max(interfaceNormals(nodeListi,i).magnitude(),1e-30);
-     }
-    
-   }
-
-}
-
 
 
 }
