+++ src/FSISPH/computeSurfaceNormalsInst.cc.py
@@ -1,0 +1,17 @@
+text = """
+//------------------------------------------------------------------------------
+// Explicit instantiation.
+//------------------------------------------------------------------------------
+#include "FSISPH/computeSurfaceNormals.cc"
+#include "Geometry/Dimension.hh"
+
+namespace Spheral {
+  template void computeSurfaceNormals(const ConnectivityMap<Dim< %(ndim)s > >&, 
+                                      const TableKernel<Dim< %(ndim)s > >&, 
+                                      const FieldList<Dim< %(ndim)s >, Dim< %(ndim)s >::Vector>&,
+                                      const FieldList<Dim< %(ndim)s >, Dim< %(ndim)s >::Scalar>&,
+                                      const FieldList<Dim< %(ndim)s >, Dim< %(ndim)s >::Scalar>&,
+                                      const FieldList<Dim< %(ndim)s >, Dim< %(ndim)s >::SymTensor>&,
+                                            FieldList<Dim< %(ndim)s >, Dim< %(ndim)s >::Vector>&);
+}
+"""
