+++ src/Pybind11Wraps/FSISPH/CMakeLists.txt
@@ -1,0 +1,1 @@
+spheral_add_pybind11_library(FSISPH)
