+++ src/Pybind11Wraps/FSISPH/FSISPHHydroBaseRZ.py
@@ -1,14 +1,13 @@
 #-------------------------------------------------------------------------------
-# SolidFSISPHHydroBase
+# SPHHydroBaseRZ
 #-------------------------------------------------------------------------------
 from PYB11Generator import *
-from SolidSPHHydroBase import *
-from RestartMethods import *
+from FSISPHHydroBase import *
 
-@PYB11template("Dimension")
-@PYB11module("SpheralSPH")
-class SolidFSISPHHydroBase(SolidSPHHydroBase):
-    "SolidFSISPHHydroBase -- SolidSPHHydro modified for large density discontinuities"
+@PYB11template()            # Override the fact SPHHydroBase is templated
+@PYB11template_dict({"Dimension" : "Dim<2>"})
+@PYB11module("SpheralFSISPH")
+class FSISPHHydroBaseRZ(FSISPHHydroBase):
 
     PYB11typedefs = """
   typedef typename %(Dimension)s::Scalar Scalar;
@@ -39,15 +38,31 @@ class SolidFSISPHHydroBase(SolidSPHHydroBase):
                HUpdate = "const HEvolutionType",
                epsTensile = "const double",
                nTensile = "const double",
-               damageRelieveRubble = "const bool",
-               negativePressureInDamage = "const bool",
-               strengthInDamage = "const bool",
                xmin = "const Vector&",
                xmax = "const Vector&"):
-        "SolidFSISPHHydroBase constructor"
+        "FSISPHHydroBaseRZ constructor"
 
     #...........................................................................
     # Virtual methods
+    @PYB11virtual
+    def initializeProblemStartup(dataBase = "DataBase<%(Dimension)s>&"):
+        "Tasks we do once on problem startup."
+        return "void"
+
+    @PYB11virtual 
+    def registerState(dataBase = "DataBase<%(Dimension)s>&",
+                      state = "State<%(Dimension)s>&"):
+        "Register the state Hydro expects to use and evolve."
+        return "void"
+
+    @PYB11virtual
+    def preStepInitialize(self,
+                          dataBase = "const DataBase<%(Dimension)s>&", 
+                          state = "State<%(Dimension)s>&",
+                          derivs = "StateDerivatives<%(Dimension)s>&"):
+        "Optional hook to be called at the beginning of a time step."
+        return "void"
+
     @PYB11virtual
     @PYB11const
     def evaluateDerivatives(time = "const Scalar",
@@ -59,26 +74,23 @@ class SolidFSISPHHydroBase(SolidSPHHydroBase):
 mass density, velocity, and specific thermal energy."""
         return "void"
 
+    @PYB11virtual
+    def applyGhostBoundaries(state = "State<%(Dimension)s>&",
+                             derivs = "StateDerivatives<%(Dimension)s>&"):
+        "Apply boundary conditions to the physics specific fields."
+        return "void"
 
-    #...........................................................................
-    # Properties
-    surfaceNormals = PYB11property("const FieldList<%(Dimension)s, Vector>&", "surfaceNormals", returnpolicy="reference_internal")
-    surfaceForceCoefficient = PYB11property("double", "surfaceForceCoefficient", "surfaceForceCoefficient",
-                           doc="additional force between different materials ala Monaghan 2013.")
-    
-    densityStabilizationCoefficient = PYB11property("double", "densityStabilizationCoefficient", "densityStabilizationCoefficient", 
-                                          doc="coefficient used to adjust velocity gradient to prevent unstable rho.")
-    
-    densityDiffusionCoefficient = PYB11property("double", "densityDiffusionCoefficient", "densityDiffusionCoefficient", 
-                                          doc="coefficient used to diffuse density amongst like nodes.")
-    
-    specificThermalEnergyDiffusionCoefficient = PYB11property("double", "specificThermalEnergyDiffusionCoefficient", "specificThermalEnergyDiffusionCoefficient", 
-                                          doc="coefficient used to diffuse specificThermalEnergy amongst like nodes.")
-    
-    sumDensityNodeLists = PYB11property("std::vector<int>", "sumDensityNodeLists", "sumDensityNodeLists", 
-                                              doc="control if rigorous density sum is applied to individual node lists.")
-    
+    @PYB11virtual
+    def enforceBoundaries(state = "State<%(Dimension)s>&",
+                          derivs = "StateDerivatives<%(Dimension)s>&"):
+        "Enforce boundary conditions for the physics specific fields."
+        return "void"
+
+    @PYB11virtual
+    @PYB11const
+    def label(self):
+        return "std::string"
 #-------------------------------------------------------------------------------
 # Inject methods
 #-------------------------------------------------------------------------------
-PYB11inject(RestartMethods, SolidFSISPHHydroBase)
+PYB11inject(RestartMethods, FSISPHHydroBaseRZ)
